feat: add incomes components

+++ src/app/main/modal-dialog/modal-dialog.component.ts
@@ -1,5 +1,6 @@
 import { Component, Input, Output, EventEmitter } from '@angular/core';
 import { NewField } from '../field.model';
+import { NewIncome } from '../income.model';
 import { DataService } from '../../services/data.service';
 
 @Component({
@@ -10,17 +11,29 @@ import { DataService } from '../../services/data.service';
 
 export class ModalDialogComponent {
   @Input() fieldDelete: NewField;
-  @Output() deleteItem = new EventEmitter<boolean>()
+  @Input() fieldIncomeDelete: NewIncome;
+
+  @Output() deleteItem = new EventEmitter<boolean>();
+  @Output() deleteItemIcome = new EventEmitter<boolean>()
 
   constructor(private dataService: DataService) { }
 
   cancel(): void {
-    this.deleteItem.emit(false);
+    if(this.fieldDelete) this.deleteItem.emit(false);
+    if(this.fieldIncomeDelete) this.deleteItemIcome.emit(false);
   }
 
   deleteId(): void {
-    this.dataService.deleteId(this.fieldDelete).subscribe();
-    this.deleteItem.emit(true);
+    if(this.fieldDelete) {
+      this.dataService.deleteId(this.fieldDelete).subscribe();
+      this.deleteItem.emit(true);
+    }
+
+    if(this.fieldIncomeDelete) {
+      this.dataService.deleteIncomeId(this.fieldIncomeDelete).subscribe();
+      this.deleteItemIcome.emit(true);
+    }
+
   }
